Widen per-root display item counter to uint16_t
The root_item_number_subitems field was uint8_t but could be
incremented beyond 255 for transactions with many displayable
leaves. Widening to uint16_t ensures the existing bounds check
can properly reject oversized root sections.

+++ app/src/tx_display.c
@@ -98,7 +98,7 @@ typedef struct {
   // total items
   uint16_t total_item_count;
   // number of items the root_item contains
-  uint8_t root_item_number_subitems[NUM_REQUIRED_ROOT_PAGES];
+  uint16_t root_item_number_subitems[NUM_REQUIRED_ROOT_PAGES];
 
   uint8_t is_default_chain;
 } display_cache_t;
